# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "NyoStore",
    "description" : "This is a NyoStore server. Here you can find \nout all the rest api's.\n",
    "version" : "2.1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/Developmenttest/NyoStore/2.1",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://authlib.nyotek.com:3010",
    "variables" : {
      "protocol" : {
        "default" : "http",
        "enum" : [ "http", "https" ]
      },
      "port" : {
        "default" : "3010",
        "enum" : [ "3010", "3020", "3030", "3040", "3050", "3060", "3070" ]
      }
    }
  } ],
...
  "paths" : {
...
    "/store/{store_key}" : {
      "get" : {
        "tags" : [ "Aggregator" ],
        "summary" : "Get store data",
        "operationId" : "getAggregatorStore",
        "parameters" : [ {
          "name" : "store_key",
          "in" : "path",
          "description" : "Store id of store to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AggregatorStoreData"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3070",
        "description" : "Aggrigator service"
      } ]
    },
    "/store/{store_key}/branch/{branch_key}" : {
      "get" : {
        "tags" : [ "Aggregator" ],
        "summary" : "Get store data",
        "operationId" : "getAggregatorStoreBranch",
        "parameters" : [ {
          "name" : "store_key",
          "in" : "path",
          "description" : "Store id of store to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "branch_key",
          "in" : "path",
          "description" : "Branch id of branch to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AggregatorStoreBranchData"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3070",
        "description" : "Aggrigator service"
      } ]
    },
    "/nearby-branches" : {
      "get" : {
        "tags" : [ "Aggregator" ],
        "summary" : "Get nearby branches",
        "operationId" : "getNearByBranches",
        "parameters" : [ {
          "name" : "latitude",
          "in" : "query",
          "description" : "Latitude of address to return",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "longitude",
          "in" : "query",
          "description" : "Longitude of address to return",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "radius",
          "in" : "query",
          "description" : "Radius of address to return",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "maxrecords",
          "in" : "query",
          "description" : "maxrecords of address to return",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Nearby-branches"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3070",
        "description" : "Aggrigator service"
      } ]
    }
  },
  "components" : {
    "schemas" : {
...
      "Error" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "Unauthorized" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          },
          "info" : {
            "type" : "string"
          }
        }
      },
      "Forbidden" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          },
          "info" : {
            "type" : "object"
          }
        }
      },
...
      "Nearby-branches" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : true
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Nearbybranches_data"
            }
          }
        },
        "example" : {
          "data" : [ {
            "distance" : 0.8008281904610115,
            "branch_id" : "branch_id",
            "address_id" : "address_id",
            "name" : "name",
            "schedule_id" : "schedule_id",
            "disbranch_id_customtance" : "disbranch_id_customtance",
            "status" : "status"
          }, {
            "distance" : 0.8008281904610115,
            "branch_id" : "branch_id",
            "address_id" : "address_id",
            "name" : "name",
            "schedule_id" : "schedule_id",
            "disbranch_id_customtance" : "disbranch_id_customtance",
            "status" : "status"
          } ],
          "message" : "message",
          "status" : true
        },
        "xml" : {
          "name" : "Nearby-branches"
        }
      },
...
      "AggregatorStoreData" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : true
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AggregatorStoreData_data"
          }
        },
        "example" : {
          "data" : {
            "store_id" : "store_id",
            "website" : "website",
            "address" : {
              "zipcode" : "zipcode",
              "country" : "country",
              "deleted" : false,
              "addr2" : "addr2",
              "address_type" : "store",
              "addr1" : "addr1",
              "city" : "city",
              "geog" : {
                "crs" : {
                  "type" : "type",
                  "properties" : {
                    "name" : "name"
                  }
                },
                "coordinates" : [ 0.8008281904610115, 0.8008281904610115 ],
                "type" : "type"
              },
              "contact" : {
                "contact_person" : "contact_person",
                "phone" : "phone",
                "contact_id" : "contact_id",
                "email" : "email"
              },
              "address_id" : "address_id",
              "state" : "state"
            },
            "documents" : [ "documents", "documents" ],
            "registration_number" : "registration_number",
            "name" : "name",
            "logo" : "logo",
            "industry" : "industry",
            "slogan" : "slogan",
            "tax_id" : "tax_id"
          },
          "message" : "message",
          "status" : true
        },
        "xml" : {
          "name" : "AggregatorStoreData"
        }
      },
      "AggregatorStoreBranchData" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : true
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AggregatorStoreBranchData_data"
          }
        },
        "example" : {
          "data" : {
            "schedule" : {
              "policy_id" : "policy_id",
              "timezone" : "timezone",
              "name" : "name",
              "schedule_id" : "schedule_id",
              "timeslots" : "timeslots"
            },
            "address" : {
              "zipcode" : "zipcode",
              "country" : "country",
              "deleted" : false,
              "addr2" : "addr2",
              "address_type" : "store",
              "addr1" : "addr1",
              "city" : "city",
              "geog" : {
                "crs" : {
                  "type" : "type",
                  "properties" : {
                    "name" : "name"
                  }
                },
                "coordinates" : [ 0.8008281904610115, 0.8008281904610115 ],
                "type" : "type"
              },
              "contact" : {
                "contact_person" : "contact_person",
                "phone" : "phone",
                "contact_id" : "contact_id",
                "email" : "email"
              },
              "address_id" : "address_id",
              "state" : "state"
            },
            "branch_id_custom" : "branch_id_custom",
            "branch_id" : "branch_id",
            "name" : "name",
            "status" : "status"
          },
          "message" : "message",
          "status" : true
        },
        "xml" : {
          "name" : "AggregatorStoreBranchData"
        }
      },
...
      "Nearbybranches_data" : {
        "type" : "object",
        "properties" : {
          "address_id" : {
            "type" : "string"
          },
          "distance" : {
            "type" : "number"
          },
          "branch_id" : {
            "type" : "string"
          },
          "disbranch_id_customtance" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "schedule_id" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "example" : {
          "distance" : 0.8008281904610115,
          "branch_id" : "branch_id",
          "address_id" : "address_id",
          "name" : "name",
          "schedule_id" : "schedule_id",
          "disbranch_id_customtance" : "disbranch_id_customtance",
          "status" : "status"
        }
      },
...
      "AggregatorStoreData_data" : {
        "type" : "object",
        "properties" : {
          "store_id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "industry" : {
            "type" : "string"
          },
          "slogan" : {
            "type" : "string"
          },
          "logo" : {
            "type" : "string"
          },
          "website" : {
            "type" : "string"
          },
          "tax_id" : {
            "type" : "string"
          },
          "registration_number" : {
            "type" : "string"
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "address" : {
            "$ref" : "#/components/schemas/AggregatorStoreData_data_address"
          }
        },
        "example" : {
          "store_id" : "store_id",
          "website" : "website",
          "address" : {
            "zipcode" : "zipcode",
            "country" : "country",
            "deleted" : false,
            "addr2" : "addr2",
            "address_type" : "store",
            "addr1" : "addr1",
            "city" : "city",
            "geog" : {
              "crs" : {
                "type" : "type",
                "properties" : {
                  "name" : "name"
                }
              },
              "coordinates" : [ 0.8008281904610115, 0.8008281904610115 ],
              "type" : "type"
            },
            "contact" : {
              "contact_person" : "contact_person",
              "phone" : "phone",
              "contact_id" : "contact_id",
              "email" : "email"
            },
            "address_id" : "address_id",
            "state" : "state"
          },
          "documents" : [ "documents", "documents" ],
          "registration_number" : "registration_number",
          "name" : "name",
          "logo" : "logo",
          "industry" : "industry",
          "slogan" : "slogan",
          "tax_id" : "tax_id"
        }
      },
...
    },
...
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}